pyproject.toml
[project]
name = "python-url-shortener"
version = "0.1.0"
description = "Um encurtador de links em Python"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cassandra-driver>=3.29.3",
    "fastapi[standard]>=0.120.1",
    "hashids>=1.3.1",
    "pydantic-settings>=2.11.0",
    "redis-py-cluster>=2.1.3",
    "taskipy>=1.14.1",
]

[tool.ruff]
line-length = 79

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.taskipy.tasks]
dev = 'fastapi dev shortner/app.py'
prod = 'fastapi run shortner/app.py'

lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'

[dependency-groups]
dev = [
    "ruff>=0.14.3",
]
